% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Example Longitudinal Data}
\format{
A dataframe with following variables:
\itemize{
\item \strong{ID}: Identity for individuals
\item \strong{VISIT}: Individuals' visit index
\item \strong{time}: Time variable
\item \strong{cov1}: Covariate
\item \strong{Group}: Group memberships for individuals
\item \strong{R1}: Response variable 1
\item \strong{R2}: Response variable 2
}
}
\usage{
ex_data
}
\description{
Example longitudinal data
}
\examples{
data(ex_data)
}
\keyword{datasets}
