% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu_Code.R
\name{Generate_Simulated_Data}
\alias{Generate_Simulated_Data}
\title{Generate Simulated Data}
\usage{
Generate_Simulated_Data(
  Npat,
  beta_l = c(5.3, 0.1, 0.3, 0.4, 0.25, -1 * 10^(-4), 3 * 10^(-8)),
  sigma2_l = 0.1^2,
  beta_d = c(1, 0.2, 0.15, 0.2, 0.15),
  sigma2_d = 0.3^2,
  nu_1 = 2.5,
  nu_2 = 1.5,
  mu = -4.8,
  beta_s1 = 1,
  beta_s2 = 0.9,
  beta_s3 = -0.75,
  beta_s4 = -5,
  h0 = 5,
  s = 1.05
)
}
\arguments{
\item{Npat}{Number of patients to simulate.}

\item{beta_l}{Linear coefficient parameter in longitudinal submodel}

\item{sigma2_l}{Error parameter in longitudinal submodel}

\item{beta_d}{Linear coefficient parameter in dosing submodel}

\item{sigma2_d}{Error parameter in dosing submodel}

\item{nu_1}{Visitation intensity peak parameter in vistation submodel}

\item{nu_2}{Visitation intensity shape parameter in vistation submodel}

\item{mu}{Baseline visitation intensity parameter in vistation submodel}

\item{beta_s1}{Creatinine-associated parameter in survival submodel}

\item{beta_s2}{Dosage-associated parameter in survival submodel}

\item{beta_s3}{Toxicity-associated parameter in survival submodel}

\item{beta_s4}{Visitation-associated parameter in survival submodel}

\item{h0}{Baseline hazard parameter in survival submodel}

\item{s}{Shape parameter in survival submodel}
}
\value{
List of data. X0_inds is the baseline data for each patient, D is the dosage, Y is the longitudinal process, and Ts are the visit times,
id are the patient id's (ranging from 1 to the total number of patients), censor and surv_time are the censoring and survival times, and Npat is the total number of patients.
}
\description{
Generate Simulated Data
}
\examples{
\dontrun{
#Simulate Data
seed=203
set.seed(seed) 
data_all<-Generate_Simulated_Data(10)
}
}
