% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_func.R
\name{mcmc_joint}
\alias{mcmc_joint}
\title{MCMC function for joint longitudinal, survival, and dosage and visitation model}
\usage{
mcmc_joint(data_all, mcmc_settings, seed)
}
\arguments{
\item{data_all}{A list of complete data. X0_inds is the baseline data for each patient, D is the dosage, Y is the longitudinal process, and Ts are the visit times,
id are the patient id's (ranging from 1 to the total number of patients), censor and surv_time are the censoring and survival times, and Npat is the total number of patients.}

\item{mcmc_settings}{A list for MCMC setup. burn.in is the number of total iterations to burn, ndisplay is number of iterations per which the display message will
appear, and Niter is the total number of iterations (including burn-in iterations).}

\item{seed}{Seed for running MCMC}
}
\value{
\item{beta_l}{Linear coefficient parameter in longitudinal submodel}
  \item{sigma2_l}{Error parameter in longitudinal submodel}
  \item{beta_d}{Linear coefficient parameter in dosing submodel}
  \item{sigma2_d}{Error parameter in dosing submodel}
  \item{nu_1}{Visitation intensity peak parameter in vistation submodel}
  \item{nu_2}{Visitation intensity shape parameter in vistation submodel}
  \item{mu}{Baseline visitation intensity parameter in vistation submodel}
  \item{beta_s1}{Creatinine-associated parameter in survival submodel}
  \item{beta_s2}{Dosage-associated parameter in survival submodel}
  \item{beta_s3}{Toxicity-associated parameter in survival submodel}
  \item{beta_s4}{Visitation-associated parameter in survival submodel}
  \item{h0}{Baseline hazard parameter in survival submodel}
  \item{shape}{Shape parameter in survival submodel}
}
\description{
MCMC function for joint longitudinal, survival, and dosage and visitation model
}
\examples{
\dontrun{
data("simulated_data")
###########
#Run MCMC 
mcmc_settings=NULL 
mcmc_settings$Niter=100
mcmc_settings$burn.in=10
mcmc_settings$ndisplay=10
mcmc_settings$peak_dist='gamma'
thin=10
post_thin_iters<-seq(mcmc_settings$burn.in+thin,mcmc_settings$Niter,thin)
seed=203
set.seed(seed)
mcmc_out_joint<-mcmc_joint(simulated_data,mcmc_settings,seed)

}

}
