% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGD_functions.R
\name{SGD_run}
\alias{SGD_run}
\title{Function for calculating optimal dosage and visit parameters for a specific patient using stochastic gradient descent (SGD)}
\usage{
SGD_run(mcmc, data_all, id_i, mcmc_inds, seed, Niter_SGD = 1000)
}
\arguments{
\item{mcmc}{output from mcmc_joint or mcmc_separate}

\item{data_all}{A list of complete data. X0_inds is the baseline data for each patient, D is the dosage, Y is the longitudinal process, and Ts are the visit times,
id are the patient id's (ranging from 1 to the total number of patients). censor and death_time are the censoring and survival times. Npat is the total number of patients.}

\item{id_i}{ID of the patient to optimize}

\item{mcmc_inds}{Vector of post-thinning MCMC iterations indices}

\item{seed}{Seed.}

\item{Niter_SGD}{Number of SGD iterations.}
}
\value{
A list containing the dosage and visit parameters across all SGD iterations. beta_d is the dosage linear coefficient
parameter vector, sigma2_d is the dosage error parameter, mu is the baseline visitation intensity, 
nu_1 is the visitation intensity peak, and nu_2 is the visitation intensity shape. The optimal value for beta_d is opt_beta_d 
and similarly for the other four parameters.
}
\description{
Function for calculating optimal dosage and visit parameters for a specific patient using stochastic gradient descent (SGD)
}
\examples{
\dontrun{
data("simulated_data")
###########
#Run MCMC 
mcmc_settings=NULL 
mcmc_settings$Niter=100
mcmc_settings$burn.in=10
mcmc_settings$ndisplay=10
mcmc_settings$peak_dist='gamma'
thin=10
post_thin_iters<-seq(mcmc_settings$burn.in+thin,mcmc_settings$Niter,thin)
seed=203
set.seed(seed)
mcmc_out_joint<-mcmc_joint(simulated_data,mcmc_settings,seed)
###########
#Run SGD
id_i=5
SGD_out<-SGD_run(mcmc_out_joint,simulated_data,id_i,post_thin_iters,seed,Niter_SGD=10)

}
}
