% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu_Code.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated dataset for 500 patients.}
\format{
List of data. X0_inds is the baseline data for each patient, D is the dosage, Y is the longitudinal process, and Ts are the visit times,
id are the patient id's (ranging from 1 to the total number of patients), censor and surv_time are the censoring and survival times, and Npat is the total number of patients.
}
\usage{
simulated_data
}
\description{
Simulated dataset for 500 patients.
}
\keyword{datasets}
