% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Test_BayTetra}
\alias{summary.Test_BayTetra}
\title{Summarize Results of BayTetra Hypothesis Test}
\usage{
\method{summary}{Test_BayTetra}(object, ...)
}
\arguments{
\item{object}{An object of class "Test_BayTetra", typically the result of calling "Test_BayTetra()".}

\item{...}{Additional arguments affecting the summary produced (currently not used).}
}
\value{
The function print the object.
}
\description{
Summarize result generated by the \code{Test_BayTetra} function in a tidy way
}
\examples{
\dontrun{
test_result <- Test_BayTetra(mcmc_result,
                           v_Rsp = paste("R", 1:2,sep = ""))
summary(test_result)
}

}
\seealso{
\code{\link{Test_BayTetra}} The function that generates the "Test_BayTetra" object.
}
